Ultima actualizacion de los test, controlador y security

+++ src/main/java/co/edu/uniquindio/proyecto/dto/ProductoGetDTO.java
@@ -27,7 +27,7 @@ public class ProductoGetDTO
 
     private int unidades;
 
-    private double precio;
+    private float precio;
 
     private int codigoUsuario;
